Pipeline templates should evaluate against a full pipeline execution
I thought there were two layers of evaluation. First the template expressions evaluated against variables on pipeline instantiation. Then the pipeline expressions evaluated on pipeline execution. It turns out that they all get evaluated at pipeline execution. It is still nice to test a pipeline template in isolation without wrapping it in a pipeline, but it needs access to a whole execution context.