Trying separate dlls for java and kotlin
Previously we were adding a DSL on top of the stage graph and trying to make it usable in Kotlin and Java. Rather than having one strategy that makes compromises for both, this adds Kotlin specific extension methods and Java specific builders that share the implementation with the extensions. This lets us optimize the Kotlin and Java experience separately. It also helps keep the DSL experience out of our simple core objects so that it is easier to iterate on or replace how one writes pipelines.