Pipeline DSL
Added a pipeline builder that actually works decently for both Kotlin and Java. Yes, after the last commit where I suggested not even trying to do that. On the Kotlin side it basically enforced what we could get for free with named parameters (which we can still do). I guess removing the commas and having the varargs methods is kind of nice too. It basically gets around some of the builder related limitations of having an immutable data class. On the Java side you can build up pipeline by chaining the setters or using them in double curly’s.
I’m not sure if this approach would work as well for StageGraph. It is nice to not have a builder on the Kotlin side and always have a legit StageGraph after every operation. Chaining those methods together is kind of nice as well to show the relationships. But I could be convinced otherwise.