Remove stage builder lambdas
Rather than a lambda to set stage attributes we can just build it into a stage graph before calling .then() on it.
Remove CloudPipeline experiment, we are moving away from convenience members on Pipeline.
The names and packages are still not ideal. There are several experiments in here. Once I rewrite the other tests with this Java DSL, and remove all the unused experimental DSLs, then the namespace will open up.